refactor: change file name from dynamic-import.ts to dynamicImport.ts
