+++ src/utilities/index.ts
@@ -5,4 +5,4 @@
 export * from '@/utilities/cancelToken'
 export * from '@/utilities/auth'
 export * from '@/utilities/sleep'
-export * from '@/utilities/dynamic-import'
+export * from '@/utilities/dynamicImport'
